pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.0",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "dj-database-url>=2.3.0",
    "django-allauth>=65.7.0",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pymysql>=1.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "qrcode>=8.1",
    "redis>=5.2.1",
    "scikit-learn>=1.6.1",
    "tzdata>=2025.2",
    "opencv-python>=4.11.0.86",
    "requests>=2.32.3",
    "imageio>=2.37.0",
    "tqdm>=4.67.1",
]
